fix(types): Move shared types from endpoints-overlay to types package.
