+++ packages/endpoints-overlay/src/EndpointsOverlay.story.tsx
@@ -1,4 +1,5 @@
 import BaseMap from "@opentripplanner/base-map";
+import { UserLocationAndType } from "@opentripplanner/types";
 import React, { ReactElement } from "react";
 import { action } from "@storybook/addon-actions";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
@@ -6,7 +7,6 @@ import { Cat } from "@styled-icons/fa-solid/Cat";
 import { Dog } from "@styled-icons/fa-solid/Dog";
 
 import EndpointsOverlay from ".";
-import { UserLocationAndType } from "./types";
 
 import "../../../node_modules/leaflet/dist/leaflet.css";
